Rearrange testing folders
