+++ src/cairo/mod.rs
@@ -6,3 +6,6 @@ pub mod errors;
 pub mod execution_trace;
 pub mod register_states;
 pub mod runner;
+
+#[cfg(test)]
+pub mod tests;
