+++ src/cairo/tests/utils.rs
@@ -1,0 +1,55 @@
+use std::ops::Range;
+
+use crate::{
+    cairo::{
+        air::{generate_cairo_proof, verify_cairo_proof},
+        runner::run::{generate_prover_args, CairoVersion},
+    },
+    starks::proof::options::ProofOptions,
+};
+
+pub fn cairo0_program_path(program_name: &str) -> String {
+    const CARGO_DIR: &str = env!("CARGO_MANIFEST_DIR");
+    const CAIRO0_BASE_REL_PATH: &str = "/cairo_programs/cairo0/";
+    let program_base_path = CARGO_DIR.to_string() + CAIRO0_BASE_REL_PATH;
+    program_base_path + program_name
+}
+
+pub fn cairo1_program_path(program_name: &str) -> String {
+    const CARGO_DIR: &str = env!("CARGO_MANIFEST_DIR");
+    const CAIRO1_BASE_REL_PATH: &str = "/cairo_programs/cairo1/";
+    let program_base_path = CARGO_DIR.to_string() + CAIRO1_BASE_REL_PATH;
+    program_base_path + program_name
+}
+
+/// Loads the program in path, runs it with the Cairo VM, and makes a proof of it
+pub fn test_prove_cairo_program(
+    file_path: &str,
+    output_range: &Option<Range<u64>>,
+    proof_mode: bool,
+) {
+    let proof_options = ProofOptions::default_test_options();
+
+    let program_content = std::fs::read(file_path).unwrap();
+    let (main_trace, pub_inputs) = generate_prover_args(
+        &program_content,
+        &CairoVersion::V0,
+        output_range,
+        proof_mode,
+    )
+    .unwrap();
+    let proof = generate_cairo_proof(&main_trace, &pub_inputs, &proof_options).unwrap();
+
+    assert!(verify_cairo_proof(&proof, &pub_inputs, &proof_options));
+}
+
+/// Loads the program in path, runs it with the Cairo VM, and makes a proof of it
+pub fn test_prove_cairo1_program(file_path: &str) {
+    let proof_options = ProofOptions::default_test_options();
+    let program_content = std::fs::read(file_path).unwrap();
+    let (main_trace, pub_inputs) =
+        generate_prover_args(&program_content, &CairoVersion::V1, &None, false).unwrap();
+    let proof = generate_cairo_proof(&main_trace, &pub_inputs, &proof_options).unwrap();
+
+    assert!(verify_cairo_proof(&proof, &pub_inputs, &proof_options));
+}
