+++ src/starks/tests/mod.rs
@@ -1,0 +1,1 @@
+pub mod integration_tests;
